+++ portmaster/minecraftbedrock/minecraftbedrock/menu/msgbox/conf.lua
@@ -1,0 +1,10 @@
+function love.conf(t)
+    t.window.title = "Minecraft Bedrock"
+    t.window.width = 640
+    t.window.height = 480
+    t.window.fullscreen = true
+    t.window.fullscreentype = "desktop"
+    t.console = false
+    t.modules.physics = false
+    t.modules.joystick = true
+end